# apps/api/openapi.yaml
openapi: 3.1.0
info:
  title: Menschlichkeit Österreich – Vereinsplattform API
  version: 2.0.0
  description: |
    Vollständige REST-API der Menschlichkeit Österreich Vereinsplattform.
    Alle geschützten Endpunkte erfordern einen JWT Bearer Token (Authorization: Bearer <token>).
  contact:
    email: admin@menschlichkeit-oesterreich.at
servers:
  - url: https://api.menschlichkeit-oesterreich.at
    description: Produktion
  - url: http://localhost:8001
    description: Lokal
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /api/forum/threads:
    get:
      tags: [Forum]
      summary: Forum-Threads abrufen
      parameters:
        - name: kategorie_id
          in: query
          schema: { type: string }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
      responses:
        '200':
          description: Threadliste
    post:
      tags: [Forum]
      summary: Thread erstellen
      responses:
        '201':
          description: Erstellt
  /api/forum/threads/{thread_id}:
    get:
      tags: [Forum]
      summary: Thread abrufen
      parameters:
        - name: thread_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Thread
    put:
      tags: [Forum]
      summary: Thread aktualisieren
      parameters:
        - name: thread_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Aktualisiert
    delete:
      tags: [Forum]
      summary: Thread löschen
      parameters:
        - name: thread_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204':
          description: Gelöscht
  /api/forum/threads/{thread_id}/posts:
    get:
      tags: [Forum]
      summary: Beiträge eines Threads abrufen
      parameters:
        - name: thread_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Beitragsliste
...